refactor(navigation): change the sidebar navigation to be more faithful
