+++ src/app/+themes/stratio/colors/colors.component.html
