+++ src/app/+themes/stratio/colors/colors.module.ts
