+++ src/app/+themes/stratio/colors/colors.routing.ts
