+++ src/app/+themes/stratio/grid/grid.component.html
@@ -1,0 +1,50 @@
+<st-api-doc [doc]="apiDoc">
+   <div class="example-run">
+      <section class="container-fluid">
+         <div class="row">
+            <div class="col-lg-24 col-md-24 col-sm-24 col-xs-24" style="background-color: #eff2f3;">
+               <div class="col-lg-24 col-md-24 col-sm-24 col-xs-24">
+                  <div class="box box-container">
+                     <div class="row">
+                        <div class="col-lg-24, col-md-24 col-sm-24 col-xs-24">
+                           <p style="background-color: #A4EDFF;">24 columns</p>
+                        </div>
+                     </div>
+                  </div>
+
+                  <div class="box box-container">
+                     <div class="row">
+                        <div class="col-lg-12, col-md-12 col-sm-12 col-xs-12">
+                           <p class="box" style="background-color: #F60">12 columns</p>
+                        </div>
+                        <div class="col-lg-12, col-md-12 col-sm-12 col-xs-12">
+                           <p class="box" style="background-color: #F60">12 columns</p>
+                        </div>
+                     </div>
+                  </div>
+
+                  <div class="box box-container">
+                     <div class="row">
+                        <div class="col-lg-8, col-md-8 col-sm-8 col-xs-8">
+                           <p class="box" style="background-color: #9C0">8 columns</p>
+                        </div>
+                        <div class="col-lg-8, col-md-8 col-sm-8 col-xs-8">
+                           <p class="box" style="background-color: #9C0">8 columns</p>
+                        </div>
+                        <div class="col-lg-8, col-md-8 col-sm-8 col-xs-8">
+                           <p class="box" style="background-color: #9C0">8 columns</p>
+                        </div>
+                     </div>
+                  </div>
+               </div>
+            </div>
+         </div>
+      </section>
+   </div>
+
+   <div class="example-code">
+      <!--{{code}}-->
+      <st-load-code file="+themes/stratio/grid/example-code/sass.html.txt"></st-load-code>
+      <st-load-code file="+themes/stratio/grid/example-code/grid.html.txt"></st-load-code>
+   </div>
+</st-api-doc>
