+++ src/app/+themes/stratio/grid/grid.component.scss
