+++ src/app/+themes/stratio/grid/grid.module.ts
