+++ src/app/+themes/stratio/icons/icon/button.routing.ts
