+++ src/app/+themes/stratio/icons/icon/icon.component.scss
