+++ src/app/+themes/stratio/icons/icons.component.ts
