+++ src/app/+themes/stratio/icons/icons.model.ts
