+++ src/app/+themes/stratio/icons/icons.module.ts
