+++ src/app/+themes/stratio/icons/icons.routing.ts
