+++ src/app/+themes/stratio/icons/icons.service.ts
