+++ src/app/+themes/stratio/typography/typography.component.html
