+++ src/app/+themes/stratio/typography/typography.component.scss
