+++ src/app/+themes/stratio/typography/typography.module.ts
